PhaseIV/backend/pyproject.toml
[tool.poetry]
name = "todo-web-app-backend"
version = "0.1.0"
description = "Backend for Todo Web App"
authors = ["Developer <developer@example.com>"]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.0"
sqlmodel = "^0.0.22"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = "^2.10.1"
pydantic-settings = "^2.7.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.20"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
cryptography = "^43.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.25.0"
httpx = "^0.27.2"
black = "^24.8.0"
flake8 = "^7.1.1"
mypy = "^1.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'

[tool.flake8]
max-line-length = 88
extend-ignore = ['E203', 'W503']
exclude = ['.git', '__pycache__', 'dist', 'build', '.venv']

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true